refactor: move away `loadCustomPlugins` from `@volar/vue-language-service`
